Create Word Boggle I
https://practice.geeksforgeeks.org/problems/word-boggle4143/0/?company[]=Amazon&company[]=Amazon&problemStatus=unsolved&problemType=functional&page=1&sortBy=submissions&category[]=Backtracking&query=company[]AmazonproblemStatusunsolvedproblemTypefunctionalpage1sortBysubmissionscompany[]Amazoncategory[]Backtracking


Given a dictionary of distinct words and an M x N board where every cell has one character. Find all possible words from the dictionary that can be formed by a sequence of adjacent characters on the board. We can move to any of 8 adjacent characters, but a word should not have multiple instances of the same cell.