Create Word Boggle II
https://practice.geeksforgeeks.org/problems/72cf44eabd18006810efad06fbb623a0682acee2/0/?category[]=DFS&category[]=DFS&page=1&query=category[]DFSpage1category[]DFS#

Given a dictionary of words and an M x N board where every cell has one character. Find all possible different words from the dictionary that can be formed by a sequence of adjacent characters on the board. We can move to any of 8 adjacent characters, but a word should not have multiple instances of the same cell.