Create Distance Nearest Cell having 1
https://practice.geeksforgeeks.org/problems/distance-of-nearest-cell-having-1-1587115620/1/?problemStatus=unsolved&page=1&category[]=DFS&category[]=Backtracking&category[]=BFS&query=problemStatusunsolvedpage1category[]DFScategory[]Backtrackingcategory[]BFS#

Given a binary grid. Find the distance of nearest 1 in the grid for each cell.
The distance is calculated as |i1 – i2| + |j1 – j2|, where i1, j1 are the row number and column number of the current cell and i2, j2 are the row number and column number of the nearest cell having value 1.