Create Number of Closed Island
https://practice.geeksforgeeks.org/problems/find-number-of-closed-islands/1/?problemStatus=unsolved&page=1&category[]=BFS&query=problemStatusunsolvedpage1category[]BFS#

A closed island is known as the group of 1s that is surrounded by only 0s on all the four sides (excluding diagonals). If any 1 is at
the edges of the given matrix then it is not considered as the part of the connected island as it is not surrounded by all 0's.