Create Word Ladder I
https://practice.geeksforgeeks.org/problems/word-ladder/1/?problemStatus=unsolved&page=1&category[]=DFS&category[]=Backtracking&category[]=BFS&query=problemStatusunsolvedpage1category[]DFScategory[]Backtrackingcategory[]BFS#

Given two distinct words startWord and targetWord, and a list denoting wordList of unique words of equal lengths. Find the length of the shortest transformation sequence from startWord to targetWord.
Keep the following conditions in mind:

A word can only consist of lowercase characters.
Only one letter can be changed in each transformation.
Each transformed word must exist in the wordList including the targetWord.
startWord may or may not be part of the wordList.