Create Matrix Love
https://practice.geeksforgeeks.org/problems/rohans-love-for-matrix4723/1/?problemStatus=unsolved&page=1&category[]=Matrix&query=problemStatusunsolvedpage1category[]Matrix#

Rohan has a special love for the matrices especially for the first element of the matrix. Being good at Mathematics, he also loves to solve the different problem on the matrices. So one day he started to multiply the matrix with the original matrix.  The elements of the original matrix are given by a00=1 , a01=1, a10=1, a11=0.
Given the power of the matrix, n calculate the an and return the a10 element mod 1000000007.