Create Euler Circuit in Undirected Grapth
https://practice.geeksforgeeks.org/problems/euler-circuit-in-a-directed-graph/1/?problemStatus=unsolved&page=1&category[]=DFS&category[]=BFS&query=problemStatusunsolvedpage1category[]DFScategory[]BFS#

Eulerian Path is a path in graph that visits every edge exactly once. Eulerian Circuit is an Eulerian Path which starts and ends on the same vertex. Given number of vertices V and adjacency list adj denoting the graph. Your task is to find that their exists the Euler circuit or not.