Create Broken Blocks
https://practice.geeksforgeeks.org/problems/broken-blocks0022/1/?problemStatus=unsolved&page=2&category[]=Matrix&category[]=DFS&query=problemStatusunsolvedpage2category[]Matrixcategory[]DFS#

In the game of Broken Blocks, the player is allowed to move on m x n blocks i.e. m levels and n stone blocks on each level such that one level is vertically above the previous level (as in a staircase), with some of its stone blocks replaced by wooden blocks. The player at the start of the game is present on the ground level (which should be considered as level ‘0’ or it can be considered as level ‘-1’). The player can start from any of the blocks present on the level ‘0’ and start moving further to next levels. The player can only move to the stone-block just above to the present stone-block or diagonally to the left or to the right. The player can’t move on the same level. If the player steps on any of the wooden block (denoted by -1), he will fall off the board and die as the wood-block will not able to hold player’s weight. Each of the stone-block has some gold coins present on it (wooden blocks doesn’t have any coins on them). If at any point the player can’t move to further level due to any reason, the game ends and his present total coin score will be considered. The player’s aim is to collect as many gold coins as he can without falling off the board.

Input: matrix = {{2,5,6},{-1,3,2},{4,-1,5}}
Output: 14
Explanation: Assume 0-based indexing.The matrix 
is:
2 5 6 (level 0)
-1 3 2 (level 1)
4 -1 5 (lever 2)
The player can collect maximum number of coins 
by moving through:matrix[0][2] + matrix[1][1] 
+ matrix[2][2] = 6 + 3 + 5 = 14 coins.