Create Cointainer With Most Water
https://practice.geeksforgeeks.org/problems/container-with-most-water0535/1/?company[]=Amazon&company[]=Amazon&problemStatus=unsolved&page=5&query=company[]AmazonproblemStatusunsolvedpage5company[]Amazon#

Given N non-negative integers a1,a2,....an where each represents a point at coordinate (i, ai). N vertical lines are drawn such that the two endpoints of line i is at (i, ai) and (i,0). Find two lines, which together with x-axis forms a container, such that it contains the most water.

Input:
N = 5
a[] = {3,1,2,4,5}
Output: 12
Explanation: 5 and 3 are distance 4 apart.
So the size of the base = 4. Height of
container = min(5, 3) = 3. So total area
= 4 * 3 = 12.