Create Search Pattern
https://practice.geeksforgeeks.org/problems/search-pattern0205/1/?company[]=Microsoft&company[]=Microsoft&problemStatus=unsolved&page=5&query=company[]MicrosoftproblemStatusunsolvedpage5company[]Microsoft#

Given two strings, one is a text string and other is a pattern string. The task is to print the indexes of all the occurences of pattern string in the text string. For printing, Starting Index of a string should be taken as 1.


Example 1:

Input:
S = "batmanandrobinarebat", pat = "bat"
Output: 1 18
Explanation: The string "bat" occurs twice
in S, one starts are index 1 and the other
at index 18.