Create Negative Weight Cycle Detection__ Used_Simple_Recursive_Logic
https://practice.geeksforgeeks.org/problems/negative-weight-cycle3504/1/?problemStatus=unsolved&page=4&category[]=Dynamic%20Programming&category[]=Graph&category[]=Graph%20colouring&query=problemStatusunsolvedpage4category[]Dynamic%20Programmingcategory[]Graphcategory[]Graph%20colouring#

Given a weighted directed graph with n nodes and m edges. Nodes are labeled from 0 to n-1, the task is to check if it contains a negative weight cycle or not.
Note: edges[i] is defined as u, v and weight.
 

Example 1:

Input: n = 3, edges = {{0,1,-1},{1,2,-2},
{2,0,-3}}
Output: 1
Explanation: The graph contains negative weight
cycle as 0->1->2->0 with weight -1,-2,-3,-1.