Create is Binary Number Multiple of 3
https://practice.geeksforgeeks.org/problems/is-binary-number-multiple-of-30654/1/?company[]=Amazon&company[]=Amazon&problemStatus=unsolved&problemType=functional&page=1&sortBy=submissions&category[]=Mathematical&category[]=Bit%20Magic&category[]=Matrix&category[]=Map&query=company[]AmazonproblemStatusunsolvedproblemTypefunctionalpage1sortBysubmissionscompany[]Amazoncategory[]Mathematicalcategory[]Bit%20Magiccategory[]Matrixcategory[]Map#


Given a binary number, Find, if given binary number is a multiple of 3. The given number can be big upto 2^10000. It is recommended to finish the task using one traversal of input binary string.

Example 1:

Input: S = "011"
Output: 1
Explanation: "011" decimal equivalent
is 3, which is divisible by 3.
â€‹Example 2:

Input: S = "100"
Output: 0
Explanation: "100" decimal equivalent
is 4, which is not divisible by 3.