Create Longest Zic-Zac Subsequence
https://practice.geeksforgeeks.org/problems/longest-zig-zag-sub-sequence2656/1/?category[]=Dynamic%20Programming&category[]=Dynamic%20Programming&problemStatus=unsolved&difficulty[]=1&difficulty[]=2&page=4&query=category[]Dynamic%20ProgrammingproblemStatusunsolveddifficulty[]1difficulty[]2page4category[]Dynamic%20Programming#

Given an array nums of n positive integers. The task is to find the longest Zig-Zag subsequence problem such that all elements of this are alternating (numsi-1 < numsi > numsi+1 or numsi-1 > numsi < numsi+1).
 

Example 1:

Input: nums = {1,5,4}
Output: 3
Explanation: The entire sequenece is a 
Zig-Zag sequence.
Examplae 2:

Input: nums = {1,17,5,10,13,15,10,5,16,8}
Ooutput: 7
Explanation: There are several subsequences
that achieve this length. 
One is {1,17,10,13,10,16,8}.