Create Articulation Points
https://practice.geeksforgeeks.org/problems/articulation-point2616/1#

Given an undirected graph (not necessarily connected) with V vertices and adjacency list adj. You are required to find all the vertices removing which (and edges through it) disconnects the graph into 2 or more components, or in other words, removing which increases the number of connected components.
Note: Indexing is zero-based i.e nodes numbering from (0 to V-1). There might be loops present in the graph.

Example 1:

Input:

Output:{1,4}
Explanation: Removing the vertex 1 will
discconect the graph as-

Removing the vertex 4 will disconnect the
graph as-

 

Your Task:
You don't need to read or print anything. Your task is to complete the function articulationPoints() which takes V and adj as input parameters and returns a list containing all the vertices removing which turn the graph into two or more disconnected components in sorted order. If there are no such vertices then returns a list containing -1.

Doctor Strange:- Same Concept
https://practice.geeksforgeeks.org/problems/doctor-strange2206/1/?problemStatus=unsolved&page=1&category[]=Graph&query=problemStatusunsolvedpage1category[]Graph#