Create Geeks and FRIENDS__Graph Coloring
https://practice.geeksforgeeks.org/problems/geek-and-friends/0/?problemStatus=unsolved&page=1&category[]=Graph&query=problemStatusunsolvedpage1category[]Graph#

Geek is hosting a contest and N students are interested in it. But some of them are friends. Geek wants to make two teams such that no two friends are on the same team.  The task is to check is it possible for the geek two make two teams or not

Note:
1. All the students are numbered from 1 to N.
2. If A is a friend of B and B is a friend of C, that doesn't mean that A is friend of C

Input:
1. The first line of the input contains a single integer T denoting the number of test cases. The description of T test cases follows.
2. The first line of each test case contains two space_separated integers N and M.
3. Next M lines contain two space-separated integers u and v, represents u is a friend of v


Output: For each test case, print the "yes" if possible, otherwise print "no" (without quotes).


Constraints:
1. 1 <= T <= 100
2. 2 <= N <= 50
3. 0 <= M <= min(1000, N*(N-1)/2)